requirements.txt
yt-dlp
librosa
pandas
moviepy
numpy
opencv-python
tqdm
transformers
torch
Pillow
openai-whisper
safetensors
scipy
ImageHash
